backend/requirements.txt
Django==3.2.15
djangorestframework==3.12.4
python-dotenv==0.21.0
Pillow==9.0.0
django-filter==23.2
gunicorn==20.1.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
drf-extra-fields==3.4.0